refactor: hoapital_map.dart
