+++ lib/screen/home.dart
@@ -2,7 +2,7 @@
 
 import 'package:ahri_manager/screen/shop_list.dart';
 import 'package:ahri_manager/screen/schedule.dart';
-import 'package:ahri_manager/screen/naver_map.dart';
+import 'package:ahri_manager/screen/hospital_map.dart';
 import 'package:ahri_manager/screen/my_animal_information.dart';
 import 'package:flutter/material.dart';
 import 'package:ahri_manager/plus/user_helper.dart';
